Move pntr_assetsys integration to its own file
