--- test/pntr_tiled_test.tmx
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<map version="1.10" tiledversion="git" orientation="orthogonal" renderorder="right-down" width="16" height="10" tilewidth="32" tileheight="32" infinite="0" nextlayerid="5" nextobjectid="1">
- <tileset firstgid="1" name="Desert" tilewidth="32" tileheight="32" spacing="1" margin="1" tilecount="48" columns="8">
-  <image source="resources/tmw_desert_spacing.png" trans="ff00ff" width="265" height="199"/>
-  <tile id="37">
-   <animation>
-    <frame tileid="37" duration="500"/>
-    <frame tileid="39" duration="500"/>
-   </animation>
-  </tile>
- </tileset>
- <layer id="1" name="Desert" width="16" height="10">
-  <data encoding="csv">
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,30,
-30,30,30,30,30,30,30,30,30,30,30,30,0,0,0,30,
-30,30,30,30,30,30,30,30,30,30,30,30,0,0,0,30,
-30,30,30,30,30,30,30,30,30,30,30,30,0,0,0,30,
-30,30,30,30,30,30,30,30,30,30,30,30,0,0,0,30
-</data>
- </layer>
- <layer id="2" name="Structure" width="16" height="10">
-  <data encoding="csv">
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,1,2,2,3,0,0,0,0,0,0,0,0,0,0,0,
-0,9,10,10,11,0,0,0,0,0,0,0,0,0,0,0,
-0,17,18,18,19,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,1,2,2,2,3,
-0,0,0,0,0,0,0,0,0,0,0,9,20,18,21,11,
-0,0,0,0,0,0,0,0,0,0,0,9,11,0,9,11,
-0,0,0,0,0,0,0,0,0,0,0,9,11,0,9,11,
-0,0,0,0,0,0,0,0,0,0,0,9,11,0,9,11
-</data>
- </layer>
- <layer id="3" name="Plants" width="16" height="10">
-  <data encoding="csv">
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,39,2147483687,0,0,0,0,0,0,
-0,0,0,0,0,0,0,1610612775,0,0,2684354599,0,0,0,0,0,
-0,0,0,0,0,0,0,536870951,0,0,3758096423,0,0,0,0,0,
-0,0,0,0,0,0,0,0,1073741863,3221225511,0,0,0,0,0,0,
-0,0,38,2147483686,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,46,46,46,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
-</data>
- </layer>
- <imagelayer id="4" name="Image Layer" opacity="0.7" offsetx="148" offsety="191.333">
-  <image source="resources/logo.png" width="128" height="128"/>
- </imagelayer>
-</map>
